Move uncompress method to another file
For ease of development for now
