+++ src/uncompress.h
@@ -1,0 +1,2 @@
+void uncompress();
+
